PCQ-1493 covering Toggled and agae check both
