+++ test/data/journeys/toggledAndAgeCheckQuestions.js
@@ -1,7 +1,10 @@
 'use strict';
 
+const AgeCheck = require('app/utils/Constants');
+
 const stepList = {
-    StartPage: 'ApplicantLanguage',
+    StartPage: 'ApplicantDateOfBirth',
+    ApplicantDateOfBirth: 'ApplicantLanguage',
     ApplicantLanguage: {
         otherLanguage: 'ApplicantEnglishLevel',
         otherwise: 'ApplicantSex'
@@ -38,13 +41,31 @@ const stepList = {
 };
 
 const toggledQuestions = [
-    {stepName: 'ApplicantDisability', ftKey: 'ft_sscs_disability_stage_1', nextStepName: 'ApplicantDisabilityImplicationAreas'},
-    {stepName: 'ApplicantDisabilityImplicationAreas', ftKey: 'ft_sscs_disability_stage_2'},
+    {stepName: 'ApplicantDateOfBirth', ftKey: 'ft_enabled'},
+    {stepName: 'ApplicantLanguage', ftKey: 'ft_disabled'},
+    {stepName: 'ApplicantSex', ftKey: 'ft_enabled', nextStepName: 'ApplicantMaritalStatus'},
+    {stepName: 'ApplicantSexualOrientation', ftKey: 'ft_disabled', nextStepName: 'ApplicantEthnicGroup'}
 ];
+
+const ageCheckQuestions = {
+    [AgeCheck.GreaterThanEighteen]: [
+        {stepName: 'ApplicantSexualOrientation'},
+        {stepName: 'ApplicantMaritalStatus'},
+    ],
+    [AgeCheck.SixteenToEighteen]: [
+        {stepName: 'ApplicantSex'},
+        {stepName: 'ApplicantGenderSameAsSex'},
+    ],
+    [AgeCheck.LessThanSixteen]: [
+        {stepName: 'ApplicantDateOfBirth'},
+        {stepName: 'ApplicantLanguage'},
+    ]
+};
 
 module.exports = () => {
     return {
         stepList,
         toggledQuestions,
+        ageCheckQuestions
     };
 };
