+++ test/unit/middleware/testSetJourney.js
@@ -7,7 +7,6 @@ const defaultJourney = require('app/journeys/default');
 const probateJourney = require('app/journeys/probate');
 const toggledQuestionsJourney = require('test/data/journeys/toggledQuestions');
 const actorDefinedJourneys = rewire('test/data/journeys/actorDefinedJourneys');
-const ageCheckQuestionsJourney = require('test/data/journeys/ageCheckQuestions');
 
 describe('setJourney', () => {
     it('should set req.journey with the default journey when no form session', async () => {
@@ -185,78 +184,5 @@ describe('setJourney', () => {
 
             revert();
         });
-    });
-
-    describe('Age Check Questions Processing', () => {
-
-        it('assigns correct skip list given null skip list', async () => {
-            const req = {
-                session: {
-                    form: {
-                        serviceId: 'TEST'
-                    },
-                    ageCheck: 2
-                }
-            };
-
-            const revert = setJourney.__set__('getBaseJourney', () => {
-                return require('test/data/journeys/ageCheckQuestions');
-            });
-
-            await setJourney(req, {});
-
-            let skipList = [
-                {stepName: 'ApplicantDateOfBirth'},
-                {stepName: 'ApplicantLanguage'},
-            ];
-
-            const journey = Object.assign({}, ageCheckQuestionsJourney());
-            journey.skipList = skipList;
-
-            expect(req.session).to.deep.equal({
-                form: {
-                    serviceId: 'TEST',
-                },
-                ageCheck: 2,
-                journey: journey
-            });
-
-            req.session.ageCheck = 1;
-            await setJourney(req, {});
-
-            skipList = [
-                {stepName: 'ApplicantSex'},
-                {stepName: 'ApplicantGenderSameAsSex'},
-            ];
-            journey.skipList = skipList;
-
-            expect(req.session).to.deep.equal({
-                form: {
-                    serviceId: 'TEST',
-                },
-                ageCheck: 1,
-                journey: journey
-            });
-
-            req.session.ageCheck = 0;
-            await setJourney(req, {});
-
-            skipList = [
-                {stepName: 'ApplicantSexualOrientation'},
-                {stepName: 'ApplicantMaritalStatus'},
-            ];
-            journey.skipList = skipList;
-
-            expect(req.session).to.deep.equal({
-                form: {
-                    serviceId: 'TEST',
-                },
-                ageCheck: 0,
-                journey: journey
-            });
-
-            revert();
-        });
-
     });
 });
